+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 # ts-appconfig
 
+## [1.1.0]
+
+### Additions
+- [Issue #5] Add readme section for options
+- [Issue #2] Add option to disable overwriting process.env
+- Typescript Template v2
+
+### Fixes
+- [Issue #4] TypeError: Cannot read properties of undefined (reading 'includes') when variable is not defined
+- [Issue #3] refval.includes is not a function when referencing non-string variable
+
 ## [1.0.0] Initial Release